Check a map against the standard it is drawn to
An orienteering map is drawn to a specification which fixes its colours,
defines every symbol down to its line widths, and says what may touch
what. `validate` reads a map and reports where it departs from one:
colours mixed wrong or printed in the wrong order, symbols redrawn,
symbols turned which should face north, contours crossing, screens
combined which may not be, and symbols so close together they merge into
one mark at printing size.

Nothing here knows what ISOM is. A `Reference` carries the standard --
its own symbol set as a map to compare against, its codes, its
thresholds -- and a caller with ISSprOM passes a different one. That is
what makes a second standard a matter of new data rather than new code.

The three geometric checks take a callback and ask it as they go, so a
caller can put a limit on how long it will wait; a check that stops early
says so rather than reporting a clean map.

Checked against the TypeScript it replaces, on a real map against ISOM
2017-2: 558 findings on each side, the same count in every category, and
441 of 443 findings identical down to their wording. Of the two that
differ, one is a gap of five and a half hundredths of a millimetre which
the two round to different hundredths, and the other is this side
declining to report a difference: where the map carries a border element
of zero width and no colour, and the standard carries none, the two draw
the same line and the older check called that a modification.

`Color` keeps its inks to full precision for the same reason. A standard
writes them to three decimals, and two colours exactly five thousandths
apart are two colours -- which is not something the storage of the number
should get a vote on.

+++ src/map.rs
@@ -219,7 +219,12 @@ pub struct Color {
     /// [`rgb`](Self::rgb) where it defines the colour by RGB alone -- so the
     /// two always describe the same colour, whichever way round the file put
     /// it.
-    pub cmyk: (f32, f32, f32, f32),
+    ///
+    /// To full precision, where [`rgb`](Self::rgb) is single: this is what a
+    /// map is specified in, and checking one against a standard compares
+    /// these to the decimal the file was written to, where a rounding of its
+    /// own would decide the answer.
+    pub cmyk: (f64, f64, f64, f64),
     /// Opacity in `[0, 1]`. Below 1 the colour is blended with whatever has
     /// already been drawn under it.
     pub opacity: f64,
